moved about page to settings
